requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
openai-whisper==20250625
edge-tts==7.2.3
transformers==5.2.0
sacremoses==0.1.1
torch==2.8.0
sentencepiece==0.2.1
